--- src/AlbumName.jsx
@@ -1,3 +1,0 @@
-export const AlbumName = (props) => {
-  return <h2>album name: {props.albumName}</h2>;
-};
