+++ src/Components/AlbumName.jsx
@@ -1,0 +1,8 @@
+export const AlbumName = (props) => {
+  return (
+    <div className="album-name">
+      {" "}
+      <p>{props.albumName}</p>
+    </div>
+  );
+};
