+++ src/Components/CoverImage.jsx
@@ -1,5 +1,5 @@
 export const CoverImage = (props) => {
   console.log(props.coverImages);
   const albumImage = props.coverImages[0];
-  return <img src={albumImage.url} alt="Music album cover image"></img>;
+  return <img clasName="cover-image" src={albumImage.url} alt="Music album cover image"></img>;
 };
