+++ src/Components/Header.jsx
@@ -1,0 +1,7 @@
+export const Header = () => {
+  return (
+    <header>
+      <h1>New Albums and Singles</h1>
+    </header>
+  );
+};
