--- src/Header.jsx
@@ -1,5 +1,0 @@
-// import data from "./data.json";
-
-// export const Header = () => {
-//   return <div> New Albums and Singles</div>;
-// };
