+++ src/index.css
@@ -1,13 +1,67 @@
 :root {
   margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
-    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
-    sans-serif;
+  padding: 0;
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen", "Ubuntu", "Cantarell", "Fira Sans",
+    "Droid Sans", "Helvetica Neue", sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
 
 code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
-    monospace;
+  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New", monospace;
+}
+body {
+  background-color: #000;
+  padding: 0 30px;
+}
+
+header {
+  display: flex;
+  justify-content: center;
+  color: #fff;
+  font-size: 16px;
+}
+
+.album-grid {
+  display: grid;
+  grid-template-columns: 100%;
+  gap: 15px;
+  justify-content: space-around;
+}
+
+.album {
+  display: flex;
+  flex-direction: column;
+}
+
+.artist-name {
+  font-family: Helvetica;
+  font-size: 14px;
+  color: #a0a0a0;
+}
+
+.album-name {
+  font-family: Helvetica;
+  font-size: 14px;
+  color: #fff;
+}
+
+@media (min-width: 768px) {
+  header {
+    font-size: 25px;
+  }
+
+  .album-grid {
+    grid-template-columns: repeat(2, 50%);
+  }
+}
+
+@media (min-width: 991px) {
+  header {
+    font-size: 35px;
+  }
+
+  .album-grid {
+    grid-template-columns: repeat(4, 25%);
+  }
 }
